Moved all of the random generation to one file
